+++ pg_documentdb/include/metadata/metadata_cache.h
@@ -214,8 +214,9 @@ Oid Float8MultiplyOperatorId(void);
 
 /* Vector Functions */
 Oid PgDoubleToVectorFunctionOid(void);
+Oid PgDoubleToSparseVecFunctionOid(bool missingOK);
 Oid VectorAsVectorFunctionOid(void);
-Oid VectorAsHalfVecFunctionOid(void);
+Oid VectorAsHalfVecFunctionOid(bool missingOK);
 Oid ApiCatalogBsonExtractVectorFunctionId(void);
 Oid ApiBsonSearchParamFunctionId(void);
 Oid ApiBsonDocumentAddScoreFieldFunctionId(void);