+++ pg_documentdb/include/vector/vector_common.h
@@ -11,6 +11,7 @@
 #define VECTOR_COMMON__H
 
 /* pgvector VECTOR_MAX_DIM: 16000 */
+/* dimensions for type vector cannot exceed 16000 */
 #define VECTOR_MAX_DIMENSIONS 16000
 
 /* ivfflat parameters */
@@ -44,6 +45,10 @@
 #define VECTOR_PARAMETER_NAME_HNSW_EF_SEARCH "efSearch"
 #define VECTOR_PARAMETER_NAME_HNSW_EF_SEARCH_STR_LEN 8
 
+/* Search parameter name for iterative scan mode */
+#define VECTOR_PARAMETER_NAME_ITERATIVE_SCAN "iterativeScan"
+#define VECTOR_PARAMETER_NAME_ITERATIVE_SCAN_STR_LEN 13
+
 /* dynamic calculation of nprobes or efSearch depending on collection size */
 #define VECTOR_SEARCH_SMALL_COLLECTION_ROWS 10000
 #define VECTOR_SEARCH_1M_COLLECTION_ROWS 1000000
@@ -68,8 +73,14 @@ extern bool EnableVectorPreFilter;
 extern bool EnableVectorPreFilterV2;
 
 /*
+ * GUC to set the iterative scan mode for pre-filtering
+ */
+extern int VectorPreFilterIterativeScanMode;
+
+/*
  * GUC to enable vector compression feature for vector search.
  */
 extern bool EnableVectorCompressionHalf;
+extern bool EnableVectorCompressionPQ;
 
 #endif