+++ pg_documentdb/include/vector/vector_spec.h
@@ -163,17 +163,18 @@ typedef Oid (*GetIndexAccessMethodOidFunc)(void);
 typedef void (*SetSearchParametersToGUCFunc)(const pgbson *searchParamBson);
 
 /*
- * Get the default SearchParamBson for the vector index.
- */
-typedef pgbson *(*GetDefaultSearchParamBsonFunc)(void);
-
-/*
  * Dynamic calculation of search parameters
  * based on the number of rows and index options.
  */
-typedef pgbson *(*CalculateSearchParamBsonFunc)(bytea *indexOptions, Cardinality
-												indexRows);
+typedef pgbson *(*CalculateSearchParamBsonFunc)(bytea *indexOptions,
+												Cardinality indexRows,
+												pgbson *searchParamBson);
 
+/*
+ * Extract the index compression type from the index options.
+ */
+typedef VectorIndexCompressionType (*ExtractIndexCompressionTypeFunc)(
+	bytea *indexOptions);
 
 /*
  * Definition of an extensible vector index.
@@ -196,8 +197,6 @@ typedef struct VectorIndexDefinition
 
 	const char *indexAccessMethodName;
 
-	bool needsReorderAfterFilter;
-
 	ParseIndexCreationSpecFunc parseIndexCreationSpecFunc;
 
 	GenerateIndexParamStringFunc generateIndexParamStrFunc;
@@ -208,9 +207,9 @@ typedef struct VectorIndexDefinition
 
 	SetSearchParametersToGUCFunc setSearchParametersToGUCFunc;
 
-	GetDefaultSearchParamBsonFunc getDefaultSearchParamBsonFunc;
-
 	CalculateSearchParamBsonFunc calculateSearchParamBsonFunc;
+
+	ExtractIndexCompressionTypeFunc extractIndexCompressionTypeFunc;
 } VectorIndexDefinition;
 
 const VectorIndexDefinition * GetVectorIndexDefinitionByIndexAmOid(Oid indexAmOid);