+++ pg_documentdb/src/configs/feature_flag_configs.c
@@ -23,7 +23,7 @@
 bool EnableVectorHNSWIndex = DEFAULT_ENABLE_VECTOR_HNSW_INDEX;
 
 /* GUC to enable vector pre-filtering feature for vector search. */
-#define DEFAULT_ENABLE_VECTOR_PRE_FILTER false
+#define DEFAULT_ENABLE_VECTOR_PRE_FILTER true
 bool EnableVectorPreFilter = DEFAULT_ENABLE_VECTOR_PRE_FILTER;
 
 #define DEFAULT_ENABLE_VECTOR_PRE_FILTER_V2 false
@@ -36,6 +36,8 @@ bool EnableVectorForceIndexPushdown = DEFAULT_ENABLE_VECTOR_FORCE_INDEX_PUSHDOWN
 #define DEFAULT_ENABLE_VECTOR_COMPRESSION_HALF true
 bool EnableVectorCompressionHalf = DEFAULT_ENABLE_VECTOR_COMPRESSION_HALF;
 
+#define DEFAULT_ENABLE_VECTOR_COMPRESSION_PQ true
+bool EnableVectorCompressionPQ = DEFAULT_ENABLE_VECTOR_COMPRESSION_PQ;
 
 /*
  * SECTION: Indexing feature flags
@@ -99,6 +101,8 @@ bool EnableMatchWithLetInLookup =
 #define DEFAULT_ENABLE_PRIMARY_KEY_CURSOR_SCAN false
 bool EnablePrimaryKeyCursorScan = DEFAULT_ENABLE_PRIMARY_KEY_CURSOR_SCAN;
 
+#define DEFAULT_USE_RAW_EXECUTOR_FOR_QUERY_PLAN true
+bool UseRawExecutorForQueryPlan = DEFAULT_USE_RAW_EXECUTOR_FOR_QUERY_PLAN;
 
 /*
  * SECTION: Top level feature flags
@@ -114,8 +118,6 @@ bool EnableBypassDocumentValidation =
 #define DEFAULT_ENABLE_NATIVE_TABLE_COLOCATION false
 bool EnableNativeTableColocation = DEFAULT_ENABLE_NATIVE_TABLE_COLOCATION;
 
-#define DEFAULT_ENABLE_USER_CRUD false
-bool EnableUserCrud = DEFAULT_ENABLE_USER_CRUD;
 
 /*
  * SECTION: Collation feature flags
@@ -149,6 +151,9 @@ bool EnableBackendStatementTimeout = DEFAULT_ENABLE_STATEMENT_TIMEOUT;
 
 #define ALTER_CREATION_TIME_IN_COMPLETE_UPGRADE false
 bool AlterCreationTimeInCompleteUpgrade = ALTER_CREATION_TIME_IN_COMPLETE_UPGRADE;
+
+#define DEFAULT_ENABLE_USERNAME_PASSWORD_CONSTRAINTS true
+bool EnableUsernamePasswordConstraints = DEFAULT_ENABLE_USERNAME_PASSWORD_CONSTRAINTS;
 
 /* FEATURE FLAGS END */
 
@@ -186,8 +191,16 @@ InitializeFeatureFlagConfigurations(const char *prefix, const char *newGucPrefix
 	DefineCustomBoolVariable(
 		psprintf("%s.enableVectorCompressionHalf", newGucPrefix),
 		gettext_noop(
-			"Enables support for vector index compression"),
+			"Enables support for vector index compression half"),
 		NULL, &EnableVectorCompressionHalf, DEFAULT_ENABLE_VECTOR_COMPRESSION_HALF,
+		PGC_USERSET, 0, NULL, NULL, NULL);
+
+	DefineCustomBoolVariable
+	(
+		psprintf("%s.enableVectorCompressionPQ", newGucPrefix),
+		gettext_noop(
+			"Enables support for vector index compression product quantization"),
+		NULL, &EnableVectorCompressionPQ, DEFAULT_ENABLE_VECTOR_COMPRESSION_PQ,
 		PGC_USERSET, 0, NULL, NULL, NULL);
 
 	DefineCustomBoolVariable(
@@ -337,13 +350,6 @@ InitializeFeatureFlagConfigurations(const char *prefix, const char *newGucPrefix
 		PGC_USERSET, 0, NULL, NULL, NULL);
 
 	DefineCustomBoolVariable(
-		psprintf("%s.enableUserCrud", newGucPrefix),
-		gettext_noop(
-			"Enables user crud through the data plane."),
-		NULL, &EnableUserCrud, DEFAULT_ENABLE_USER_CRUD,
-		PGC_USERSET, 0, NULL, NULL, NULL);
-
-	DefineCustomBoolVariable(
 		psprintf("%s.enableNowSystemVariable", newGucPrefix),
 		gettext_noop(
 			"Enables support for the $$NOW time system variable."),
@@ -407,6 +413,14 @@ InitializeFeatureFlagConfigurations(const char *prefix, const char *newGucPrefix
 		PGC_USERSET, 0, NULL, NULL, NULL);
 
 	DefineCustomBoolVariable(
+		psprintf("%s.useRawExecutorForQueryPlan", newGucPrefix),
+		gettext_noop(
+			"Whether or not to enable using the raw executor for query plans."),
+		NULL, &UseRawExecutorForQueryPlan,
+		DEFAULT_USE_RAW_EXECUTOR_FOR_QUERY_PLAN,
+		PGC_USERSET, 0, NULL, NULL, NULL);
+
+	DefineCustomBoolVariable(
 		psprintf("%s.alterCreationTimeInCompleteUpgrade", newGucPrefix),
 		gettext_noop(
 			"alter creation_time of data table."),
@@ -420,5 +434,13 @@ InitializeFeatureFlagConfigurations(const char *prefix, const char *newGucPrefix
 			"use the unsafe transform for index term elements."),
 		NULL, &IndexTermUseUnsafeTransform,
 		DEFAULT_USE_UNSAFE_INDEX_TERM_TRANSFORM,
+		PGC_USERSET, 0, NULL, NULL, NULL);
+
+	DefineCustomBoolVariable(
+		psprintf("%s.enableUsernamePasswordConstraints", newGucPrefix),
+		gettext_noop(
+			"Determines whether username and password constraints are enabled."),
+		NULL, &EnableUsernamePasswordConstraints,
+		DEFAULT_ENABLE_USERNAME_PASSWORD_CONSTRAINTS,
 		PGC_USERSET, 0, NULL, NULL, NULL);
 }