+++ pg_documentdb/src/metadata/metadata_cache.c
@@ -598,6 +598,9 @@ typedef struct DocumentDBApiOidCacheData
 	/* OID of the array_to_vector function. */
 	Oid PgDoubleToVectorFunctionOid;
 
+	/* OID of the array_to_sparsevec function. */
+	Oid PgDoubleToSparseVecFunctionOid;
+
 	/* OID of the vector as vector Cast function */
 	Oid VectorAsVectorFunctionOid;
 
@@ -2777,6 +2780,32 @@ PgDoubleToVectorFunctionOid(void)
 
 
 /*
+ * Returns the function Oid for converting a double[] to a sparse vector specifically the array_to_sparsevec function.
+ * Note: the array_to_sparsevec function is introduced in pgvector 0.8.0,
+ * Currently it is only used for checking if the version of pgvector is 0.8.0
+ * Set missingOK to true, this function will return InvalidOid if the pgvector version is lower than that.
+ */
+Oid
+PgDoubleToSparseVecFunctionOid(bool missingOK)
+{
+	InitializeDocumentDBApiExtensionCache();
+
+	if (Cache.PgDoubleToSparseVecFunctionOid == InvalidOid)
+	{
+		List *functionNameList = list_make2(makeString("public"),
+											makeString("array_to_sparsevec"));
+
+		Oid paramOids[3] = { FLOAT8ARRAYOID, INT4OID, BOOLOID };
+
+		Cache.PgDoubleToSparseVecFunctionOid =
+			LookupFuncName(functionNameList, 3, paramOids, missingOK);
+	}
+
+	return Cache.PgDoubleToSparseVecFunctionOid;
+}
+
+
+/*
  * VectorAsVectorFunctionOid returns the OID of the vector as vector cast function.
  */
 Oid
@@ -2801,9 +2830,11 @@ VectorAsVectorFunctionOid(void)
 
 /*
  * VectorAsHalfVecFunctionOid returns the OID of the vector as half vector cast function.
+ * Note: with older versions of pgvector, the cast function "public.vector_to_halfvec" is not supported.
+ * So we need to check InvalidOid
  */
 Oid
-VectorAsHalfVecFunctionOid(void)
+VectorAsHalfVecFunctionOid(bool missingOK)
 {
 	InitializeDocumentDBApiExtensionCache();
 
@@ -2813,7 +2844,6 @@ VectorAsHalfVecFunctionOid(void)
 											makeString("vector_to_halfvec"));
 
 		Oid paramOids[3] = { VectorTypeId(), INT4OID, BOOLOID };
-		bool missingOK = false;
 		Cache.VectorAsHalfVecFunctionOid =
 			LookupFuncName(functionNameList, 3, paramOids, missingOK);
 	}
@@ -3276,11 +3306,13 @@ Oid
 BsonDollaMergeDocumentsFunctionOid(void)
 {
 	bool missingOk = false;
-	return GetDocumentDBInternalBinaryOperatorFunctionId(
+	int nargs = 3;
+	Oid argTypes[3] = { BsonTypeId(), BsonTypeId(), BOOLOID };
+	return GetSchemaFunctionIdWithNargs(
 		&Cache.ApiInternalSchemaBsonDollarMergeDocumentsFunctionOid,
+		DocumentDBApiInternalSchemaName,
 		"bson_dollar_merge_documents",
-		BsonTypeId(),
-		BsonTypeId(), missingOk);
+		nargs, argTypes, missingOk);
 }