+++ pg_documentdb/src/test/regress/sql/user_crud_commands.sql
@@ -1,8 +1,7 @@
 SET documentdb.next_collection_id TO 1972800;
 SET documentdb.next_collection_index_id TO 1972800;
 
---set Feature flag for user crud
-SET documentdb.enableUserCrud TO ON;
+SET documentdb.maxUserLimit TO 10;
 \set VERBOSITY TERSE
 
 show documentdb.blockedRolePrefixList;
@@ -54,6 +53,17 @@ SELECT documentdb_api.create_user('{"createUser":"test_user4", "pwd":"test_passw
 
 --Create a user with no DB
 SELECT documentdb_api.create_user('{"createUser":"test_user4", "pwd":"test_password", "roles":[{"role":"readWriteAnyDatabase"}, {"role":"clusterAdmin"}]}');
+
+-- Create a user with an empty password should fail
+SELECT documentdb_api.create_user('{"createUser":"test_user_empty_pwd", "pwd":"", "roles":[{"role":"readAnyDatabase","db":"admin"}]}');
+
+-- Create a user with password less than 8 characters and drop it
+SELECT documentdb_api.create_user('{"createUser":"test_user_short_pwd", "pwd":"Short1!", "roles":[{"role":"readAnyDatabase","db":"admin"}]}');
+SELECT documentdb_api.drop_user('{"dropUser":"test_user_short_pwd"}');
+
+-- Create a user with password more than 256 characters and drop it
+SELECT documentdb_api.create_user('{"createUser":"test_user_long_pwd", "pwd":"ThisIsAVeryLongPasswordThatExceedsTheTwoHundredFiftySixCharacterLimitAndThereforeShouldFailValidation1!abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789", "roles":[{"role":"readAnyDatabase","db":"admin"}]}');
+SELECT documentdb_api.drop_user('{"dropUser":"test_user_long_pwd"}');
 
 --Verify that the user is created
 SELECT documentdb_api.users_info('{"usersInfo":"test_user4"}');
@@ -123,9 +133,6 @@ SELECT current_user as original_user \gset
 -- switch to read only user
 \c regression readOnlyUser
 
---set Feature flag for user crud
-SET documentdb.enableUserCrud TO ON;
-
 --Create without privileges
 SELECT documentdb_api.create_user('{"createUser":"newUser", "pwd":"test_password", "roles":[{"role":"readAnyDatabase","db":"admin"}]}');
 
@@ -134,9 +141,6 @@ SELECT documentdb_api.drop_user('{"dropUser":"test_user"}');
 
 -- switch to admin user
 \c regression adminUser
-
---set Feature flag for user crud
-SET documentdb.enableUserCrud TO ON;
 
 --Create without privileges
 SELECT documentdb_api.create_user('{"createUser":"newUser", "pwd":"test_password", "roles":[{"role":"readAnyDatabase","db":"admin"}]}');
@@ -200,5 +204,5 @@ SELECT documentdb_api.drop_user('{"dropUser":"test_user_injection_attack"}');
 SELECT documentdb_api.drop_user('{"dropUser":"readOnlyUser"}');
 SELECT documentdb_api.drop_user('{"dropUser":"adminUser"}');
 
--- Reset the max user limit to 10
-SET documentdb.maxUserLimit TO 10;
+-- Reset the max user limit to 500
+SET documentdb.maxUserLimit TO 500;