+++ pg_documentdb/src/vector/vector_utilities.c
@@ -25,9 +25,11 @@
 #include "metadata/metadata_cache.h"
 #include "vector/bson_extract_vector.h"
 #include "vector/vector_common.h"
+#include "vector/vector_configs.h"
 #include "vector/vector_planner.h"
 #include "vector/vector_utilities.h"
 #include "vector/vector_spec.h"
+#include "utils/error_utils.h"
 
 
 /* --------------------------------------------------------- */
@@ -44,6 +46,9 @@ static Expr * GenerateVectorExractionExprFromQueryWithCast(Node *vectorQuerySpec
 static VectorIndexDistanceMetric GetDistanceMetricFromOpId(Oid similaritySearchOpId);
 static VectorIndexDistanceMetric GetDistanceMetricFromOpName(const
 															 char *similaritySearchOpName);
+
+static bool IsHalfVectorCastFunctionCore(FuncExpr *vectorCastFunc,
+										 bool logWarning);
 
 /* --------------------------------------------------------- */
 /* Top level exports */
@@ -96,11 +101,10 @@ EvaluateMetaSearchScore(pgbson *document)
  *  b. If the number of rows is greater than 10K, the default efSearch is HNSW_DEFAULT_EF_SEARCH.
  */
 pgbson *
-CalculateSearchParamBsonForIndexPath(IndexPath *vectorSearchPath)
+CalculateSearchParamBsonForIndexPath(IndexPath *vectorSearchPath, pgbson *searchParamBson)
 {
 	IndexPath *indexPath = (IndexPath *) vectorSearchPath;
 	Oid indexRelam = indexPath->indexinfo->relam;
-	pgbson *searchParamBson = NULL;
 
 	/* Get rows in the index */
 	Cardinality indexRows = indexPath->indexinfo->tuples;
@@ -118,7 +122,7 @@ CalculateSearchParamBsonForIndexPath(IndexPath *vectorSearchPath)
 		if (indexRelation->rd_options != NULL)
 		{
 			searchParamBson = definition->calculateSearchParamBsonFunc(
-				indexRelation->rd_options, indexRows);
+				indexRelation->rd_options, indexRows, searchParamBson);
 		}
 
 		RelationClose(indexRelation);
@@ -217,6 +221,16 @@ GenerateVectorIndexExprStr(const char *keyPath,
 
 
 /*
+ * Checks if the vector cast function is a cast to half vector.
+ */
+bool
+IsHalfVectorCastFunction(FuncExpr *vectorCastFunc)
+{
+	return IsHalfVectorCastFunctionCore(vectorCastFunc, false);
+}
+
+
+/*
  * Checks if a query path matches a vector index and returns the index
  * expression function of the vector index.
  */
@@ -253,8 +267,9 @@ IsMatchingVectorIndex(Relation indexRelation, const char *queryVectorPath,
 	}
 
 	FuncExpr *vectorCtrExpr = (FuncExpr *) linitial(indexprs);
+	bool logWarning = true;
 	if (vectorCtrExpr->funcid != VectorAsVectorFunctionOid() &&
-		vectorCtrExpr->funcid != VectorAsHalfVecFunctionOid())
+		!IsHalfVectorCastFunctionCore(vectorCtrExpr, logWarning))
 	{
 		/* Any other index with function expression is not valid vector index */
 		return false;
@@ -307,7 +322,7 @@ GenerateVectorSortExpr(VectorSearchOptions *vectorSearchOptions,
 	/* For the exact search, we don't use the vector index
 	 * so force the cast function to the full vector if it is a half vector */
 	if (vectorSearchOptions->exactSearch &&
-		vectorCastFunc->funcid == VectorAsHalfVecFunctionOid())
+		IsHalfVectorCastFunction(vectorCastFunc))
 	{
 		/* copy the vector cast expr, change the function id to full vector */
 		vectorCastFunc = (FuncExpr *) copyObject(vectorCastFunc);
@@ -408,31 +423,6 @@ SetSearchParametersToGUC(Oid vectorAccessMethodOid, pgbson *searchParamBson)
 
 
 /*
- * This function set the default search parameters to inputState if the user did not specify
- */
-void
-TrySetDefaultSearchParamForCustomScan(SearchQueryEvalData *querySearchData)
-{
-	pgbson *searchInput = NULL;
-	if (querySearchData->SearchParamBson != (Datum) 0)
-	{
-		searchInput = DatumGetPgBson(querySearchData->SearchParamBson);
-	}
-
-	if (searchInput == NULL || IsPgbsonEmptyDocument(searchInput))
-	{
-		const VectorIndexDefinition *definition = GetVectorIndexDefinitionByIndexAmOid(
-			querySearchData->VectorAccessMethodOid);
-		if (definition != NULL)
-		{
-			querySearchData->SearchParamBson = PointerGetDatum(
-				definition->getDefaultSearchParamBsonFunc());
-		}
-	}
-}
-
-
-/*
  * This function returns the full vector operator id for the given distance metric.
  */
 Oid
@@ -463,6 +453,19 @@ GetFullVectorOperatorId(VectorIndexDistanceMetric distanceMetric)
 	}
 
 	return InvalidOid;
+}
+
+
+/*
+ * The pgvector iterative scan is available from 0.7.0,
+ */
+bool
+IsPgvectorHalfVectorAvailable(void)
+{
+	/* public.vector_to_halfvec function is introduced in 0.7.0 */
+	/* so we can check the function is available or not to indicate the half vector is available */
+	bool missingOk = true;
+	return OidIsValid(VectorAsHalfVecFunctionOid(missingOk));
 }
 
 
@@ -541,4 +544,34 @@ GetDistanceMetricFromOpName(const char *similaritySearchOpName)
 	{
 		return VectorIndexDistanceMetric_Unknown;
 	}
+}
+
+
+/*
+ * Checks if the vector cast function is a cast to half vector.
+ * This is used to check if the vector index is a half vector index.
+ */
+static bool
+IsHalfVectorCastFunctionCore(FuncExpr *vectorCastFunc, bool logWarning)
+{
+	if (!IsPgvectorHalfVectorAvailable())
+	{
+		if (logWarning)
+		{
+			ereport(WARNING, (errmsg(
+								  "The half vector is not supported by pgvector, please check the version of pgvector")));
+		}
+		return false;
+	}
+
+	bool missingOk = false;
+	Oid halfVectorCastFuncOid = VectorAsHalfVecFunctionOid(missingOk);
+
+	if (vectorCastFunc != NULL &&
+		vectorCastFunc->funcid == halfVectorCastFuncOid)
+	{
+		return true;
+	}
+
+	return false;
 }